fix(player): correcting incorrect behavior and errors
In the PlayerActivity.kt file, passing the callback function through the entire dependency graph has been removed. The removed function was replaced by the consume parameter in the preparePlayer() function of the Playable interface. Also, the consume parameter was added as a parameter in the prepare player function in the PlayerInteractor and PlayerRepository interfaces.
In the PlayerForMusic.kt file, obtaining an instance of the MediaPlayer class using Koin was added (A corresponding dependency was declared in the DataModule.kt file).