feat(media): implementation of media library screen layout and TabLayout and viewpager customization
The layout of the media library screen was implemented according to the layout.
An Adapter for viewpager was created and all necessary logic was implemented in it.
TabLayoutMediator was created and configured in MediaActivity.kt to combine TabLayout and ViewPager. The functionality of the button to return to the previous screen was also implemented here.