perf(search): implementation of debounce and search tracks request on coroutines
In SearchFragment.kt, the debounce functions for clicking on a list item and executing a search query have been rewritten from Handler to coroutines.

In the SearchViewModel.kt file, the implementation of the function for getting the track list to display on the search screen was replaced. In particular, the logic for querying Interactor on coroutines was implemented.